config-files/requirements-config.txt
mwparserfromhell
tqdm
langchain
langchain-huggingface
langchain-qdrant
python-dotenv
sentence-transformers
langchain-text-splitters
hf_xet
qdrant-client
langchain-core
